minor warnings on on win32 fixed
git-svn-id: http://svn.fastmq.org/svn/zeromq/trunk@845 7451a71c-b6c8-49cc-8254-141fc5361860

+++ perf/tests/zmq/c_local_lat.c
@@ -66,8 +66,8 @@ typedef __int64 uint64_t;
 
         /*  Convert the tick number into the number of seconds  */
         /*  since the system was started...  */
-        ticksDivM = ticksPerSecond.QuadPart / 1000000000;
-        time.QuadPart = tick.QuadPart / ticksDivM;
+        ticksDivM = (double) ticksPerSecond.QuadPart / 1000000000;
+        time.QuadPart = (ULONGLONG) (tick.QuadPart / ticksDivM);
        
       
         return time.QuadPart;
